+++ src/api_types/personal_best.rs
@@ -1,12 +1,14 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Debug)]
+use super::AttemptResult;
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
 #[serde(rename_all = "camelCase")]
 pub struct PersonalBest {
-        pub event_id: String,
-        pub best: AttemptResult,
-        pub r#type: String,
-        pub world_ranking: u64,
-        pub continental_ranking: u64,
-        pub national_ranking: u64,
+    pub event_id: String,
+    pub best: AttemptResult,
+    pub r#type: String,
+    pub world_ranking: u32,
+    pub continental_ranking: u32,
+    pub national_ranking: u32,
 }
